fix(entities-endpoints): move ICalendarRangeService out of Internal namespace
PR #1700 introduced the public `ICalendarRangeService` interface and
`CalendarRange` record struct in `Granit.Entities.Endpoints.Internal.*`.
The architecture rule
`ClassDesignTests.Public_types_should_not_reside_in_Internal_namespaces`
flagged it on every PR since: public types in `*.Internal.*` are
invisible-by-convention to consumers, but consumers DO need to implement
this interface (host-supplied EF Core executor per #1689).

Fix: split the file. Public types (the interface + the record struct)
move to `src/Granit.Entities.Endpoints/ICalendarRangeService.cs` under
the root namespace `Granit.Entities.Endpoints`. The default no-op
`NullCalendarRangeService` (internal sealed class) stays in
`Internal/NullCalendarRangeService.cs` (renamed to satisfy the
file-name-matches-type-name archi rule).

The DI registration in `EntitiesEndpointRouteBuilderExtensions.cs`
already imports `Granit.Entities.Endpoints.Internal` for
`NullCalendarRangeService`; the public namespace `Granit.Entities.Endpoints`
is reachable transitively via the existing namespace declaration. No
caller change needed.

Architecture shard: 207/207 (was 206/207 with this exact failure).
Build clean. No behavior change.
